refactor: 중첩된 코드 수정 및 구조 개선
