+++ src/components/MovieDetailModal/MovieDetailModal.style.css
