+++ src/components/MovieDetailModal/index.ts
@@ -1,7 +1,6 @@
 import { posterNotFoundImage, starFilledImage } from '../../assets/images';
-import { CLASS } from '../../constants/selector';
 import './MovieDetailModal.style.css';
-import { IMAGE_URL } from '../../constants';
+import { IMAGE_URL, NO_OVERVIEW_MESSAGE } from '../../constants';
 import StarRate from './StarRate';
 import type { MovieDetail } from '../../types/movie';
 
@@ -15,8 +14,6 @@ const MovieDetailModal = {
           <h2 id="movie-detail-title"></h2>
           <div id="movie-detail-main">
             <img
-              class="${CLASS.SKELETON}"
-              src=""
               onerror="this.src='${posterNotFoundImage}'"
               loading="lazy"
               alt=""
@@ -39,6 +36,15 @@ const MovieDetailModal = {
   },
   setEvent(target: HTMLElement) {
     StarRate.setEvent(target);
+
+    const modal = document.querySelector<HTMLDialogElement>('dialog');
+    if (modal === null) return;
+
+    const closeButton = modal.querySelector('#movie-detail-close') as HTMLButtonElement;
+    closeButton.addEventListener('click', () => modal.close());
+
+    const backdrop = modal.querySelector('.modal-backdrop') as HTMLDivElement;
+    backdrop.addEventListener('click', () => modal.close());
   },
   open(data: MovieDetail) {
     const modal = document.querySelector<HTMLDialogElement>('dialog');
@@ -51,9 +57,6 @@ const MovieDetailModal = {
     const title = modal.querySelector('h2') as HTMLHeadElement;
     title.innerHTML = data.title;
 
-    const closeButton = modal.querySelector('#movie-detail-close') as HTMLButtonElement;
-    closeButton.addEventListener('click', () => modal.close());
-
     const poster = modal.querySelector('img') as HTMLImageElement;
     poster.src = IMAGE_URL + data.poster_path;
 
@@ -64,10 +67,7 @@ const MovieDetailModal = {
     rate.innerText = data.vote_average.toFixed(1);
 
     const overview = modal.querySelector('#movie-detail-overview') as HTMLParagraphElement;
-    overview.innerText = data.overview !== '' ? data.overview : 'overview가 존재하지 않습니다.';
-
-    const backdrop = modal.querySelector('.modal-backdrop') as HTMLDivElement;
-    backdrop.addEventListener('click', () => modal.close());
+    overview.innerText = data.overview !== '' ? data.overview : NO_OVERVIEW_MESSAGE;
   },
 };
