+++ src/constants/index.ts
@@ -10,6 +10,8 @@ export const ERROR_MESSAGE: ErrorMessage = {
 
 export const SCORE_MESSAGE = ['별점매기기', '최악이에요', '별로에요', '보통이에요', '재미있어요', '명작이에요'];
 
+export const NO_OVERVIEW_MESSAGE = 'overview가 존재하지 않습니다.';
+
 export const DEFAULT_LIST_LENGTH = 20;
 
 export const MAX_PAGE = 500;
